Working script for LDR sensor
This is pretty substantial rewrite to get the sensors working correctly and fix the multiple sending issues. I completely cut out the ini file to make sure that everything was working correctly with both Pushover an IFTTT.

Changes:
-Merged ini into python script
-Cut out other services beyond Pushover and IFTTT
-Removed delays since they're not needed when using an LDR, no worries of false alarms or different machine cycles
-Only monitoring "Machine Off" event
-Trimmed variables and adjusted IF function calculations

So far I have only had to slightly adjust the board's potentiometer to cut out ambient lighting, but a simple shroud would have sufficed (unfortunately I only have black sugru and the wife said no to that).

Please incorporate any of this into the main branch. I'll be happy to test .